[ui-storagebrowser] chore: move useHuePubSub to hooks folder
